OnionSoup: add <meta name=timeout content=long> to WPT in SlowTests
This is the 2nd step in the process of banning external/wpt from
SlowTests. After cleaning up the wtp entries we're marking them as
slow by using wpt idioms, i.e.:

1) Add <meta name="timeout" content="long" /> to .html files
2) Add // META: timeout=long to .js files

Some of the tests will be uploaded in a follow up CL because they are
so big that uploading a single CL for all of them fails.

Bug: 765026
Change-Id: Iaf051104f419524916e19324996cc36ee37d5426