IndexedDB: Fix bindings-inject-key.html to use Object.defineProperty()
In Chrome, assigning to Object.property.a under chromedriver
causes a stack size overflow exception.

I believe the upstream wpt.fyi failures in this test under
Chrome are due to the same issue where that assignment in
the third test in this file causes a JS stack overflow.

Since other tests in this file use Object.defineProperty(),
I wanted to toss out a possible quick fix to the test which
updates the third test to align with the way the first two
tests configure Object.prototype.